First steps towrads default REST support
